Deduplicate Dav1dFrameContext and inner structs
